+++ src/shared/block_utils.hpp
@@ -36,7 +36,7 @@ vec3i64 wc2bc(vec3i64 wc);
 vec3i64 bc2cc(vec3i64 bc);
 vec3ui8 bc2icc(vec3i64 bc);
 
-size_t gridCycleIndex(vec3i64 v, size_t gridSize);
+int64 gridCycleIndex(vec3i64 v, int64 gridSize);
 
 size_t vec3i64HashFunc(vec3i64 v);
